test(webapp): reduce repeated container setup
